Correct routes, resource-type component & add new rest-handler component
