+++ define/public/index.html
@@ -6,6 +6,11 @@
     <!-- <link href="style/style.css" rel="stylesheet" type="text/css"> -->
     <meta name="robots" content="all" />
     <meta name="viewport" content="width=device-width" />
+    <style type="text/css">
+    html{
+    	font-family: Neris;
+    }
+    </style>
    
 </head>
 <body>
