+++ define/src/js/components/routes.jsx
@@ -9,7 +9,15 @@ var routes = (
 	<Route path="/" handler={Sealious.App}>
 		<Route path="resource-type" >
 			<DefaultRoute name="resource-type-list"/>
-			<Route name="resource-type-view" path=":resource_name" handler={Sealious.ResourceType.View}/>
+			<Route name="resource-type-view" path=":resource_name" handler={Sealious.ResourceType}>
+				<Route name="rest" path="rest" handler={Sealious.RestHandler}>
+					<Route name="get" path="get" handler={Sealious.RestHandler.get}/>
+					<Route name="post" path="post" handler={Sealious.RestHandler.post}/>
+					<Route name="put" path="put" handler={Sealious.RestHandler.put}/>
+					<Route name="patch" path="patch" handler={Sealious.RestHandler.patch}/>
+					<Route name="delete" path="delete" handler={Sealious.RestHandler.delete}/>
+				</Route>
+			</Route>
 		</Route>
 	</Route>
 );
